% Generated by roxygen2 (4.0.2): do not edit by hand
\name{makeMap}
\alias{makeMap}
\title{Creates a Map}
\usage{
makeMap(data, map, palette = "Greens", title = "", fill_label = "",
  output = "out.png", width = 7, height = 7, per = FALSE,
  reverse_colors = FALSE, reverse_legend = TRUE,
  projection = c("mercator", "albers"), proj_args = NULL, border = NULL,
  border_color = "black", size = 0.25)
}
\arguments{
\item{data}{Dataframe with the data you want to plot.  Must have
two columns with the first column called "id" matching up with the
"id" column in the map dataframe.}

\item{map}{Dataframe with shapefile data (after using fortify()).}

\item{palette}{Color Brewer palette to use (http://colorbrewer2.org/).}

\item{title}{String with title of the map, defaults to blank.}

\item{fill_label}{String with a label for a legend, defaults to blank.}

\item{output}{String with output filename, defaults to "out.png".}

\item{width}{Defaults to 7.}

\item{height}{Defaults to 7.}

\item{per}{TRUE if labels are a percentage (0.2 -> 20\%), defaults to FALSE.}

\item{reverse_colors}{TRUE if you want to reverse the colors on the legend,
defaults to FALSE.}

\item{reverse_legend}{TRUE if you want to reverse the legend, defaults to TRUE.}

\item{projection}{Method of projection, takes in Mercator or Albers.}

\item{proj_args}{Any arguments necessary for the projection method
(Albers requires lat0 and lat1 arguments).}

\item{border}{Optional parameter if you want different borders
(i.e. census tract data with county borders).}

\item{border_color}{Color of the borders, defaults to black.}

\item{size}{Size of the borders, defaults to 0.25.}
}
\description{
This function takes in a dataframe, shapefile path,
and variable and outputs a .png file
}
\examples{
makemap(data=d, map = map)
}

